backend_api/requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10.5
django-filter==1.0.2
django-webpack-loader==0.4.1
djangorestframework==3.6.2
gunicorn==19.7.0
packaging==16.8
psycopg2==2.7.1
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
whitenoise==3.3.0
